Update models namespace
