+++ src/Models/Internal/Resource/Orders/Post.php
@@ -1,12 +1,12 @@
 <?php
 
-namespace JacobDeKeizer\Ccv\Models\Resource\Orders;
+namespace JacobDeKeizer\Ccv\Models\Internal\Resource\Orders;
 
 use JacobDeKeizer\Ccv\Contracts\Model;
 use JacobDeKeizer\Ccv\Traits\FromArray;
 use JacobDeKeizer\Ccv\Traits\ToArray;
 
-class Patch implements Model
+class Post implements Model
 {
     use FromArray, ToArray;
 
@@ -16,7 +16,12 @@ class Patch implements Model
      private $invoicenumber;
 
     /**
-     * @var bool|null If the order is marked as paid.
+     * @var bool|null If false, the order won't be marked as completed. The basket_href and checkout_href will be available for this resource. The consumer can be forwarded to these urls and complete the order. By default this value is TRUE.
+     */
+     private $isCompleted;
+
+    /**
+     * @var bool|null If the order is marked as paid
      */
      private $paid;
 
@@ -26,7 +31,7 @@ class Patch implements Model
      private $safetyDepositReturned;
 
     /**
-     * @var float|null Total shipping costs. If this order already has a value for this field, it will be overwritten.
+     * @var float|null Total shipping costs
      */
      private $totalShipping;
 
@@ -36,12 +41,12 @@ class Patch implements Model
      private $paymethodCosts;
 
     /**
-     * @var float|null Extra costs added to the order, for instance as handling costs. This can be used as a discount less than zero. If this order already has a value for this property, it will be overwritten.
+     * @var float|null Extra costs added to the order, for instance as handling costs. This can be used as a discount less than zero.
      */
      private $extraCosts;
 
     /**
-     * @var string|null Description of the extra costs. This is visible on the invoice and only if extra_costs is not zero. If extra_costs is supplied, it's advised to fill this property.
+     * @var string|null Description of the extra costs. This is visible on the invoice.
      */
      private $extraCostsDescription;
 
@@ -51,12 +56,12 @@ class Patch implements Model
      private $taxesIncluded;
 
     /**
-     * @var bool|null If order row prices contain taxes. Use this property to choose between an inc. VAT order and an ex. VAT order.
+     * @var bool|null If order row prices contain taxes. Use this field to choose between an inc. VAT order and an ex. VAT order.
      */
      private $orderRowTaxesIncluded;
 
     /**
-     * @var bool|null If shippingcosts are included in the total tax amount.
+     * @var bool|null If shippingcosts are included in the total tax amount
      */
      private $shippingTaxesIncluded;
 
@@ -81,12 +86,7 @@ class Patch implements Model
      private $status;
 
     /**
-     * @var string|null When the boolean 'mail' is set, the note will be appended to the email which is send to the customer.
-     */
-     private $note;
-
-    /**
-     * @var string|null Track and Trace Code, supplied by shipping party. When the boolean 'mail' is set, the Track and Trace Code can be appended to the email which is send to the customer.
+     * @var string|null Track and Trace code, usually provided by the shipping service. If a note is emailed to the customer this value will be included.
      */
      private $trackAndTraceCode;
 
@@ -94,11 +94,6 @@ class Patch implements Model
      * @var string|null Track & Trace Carrier. This value represents the shipping service.
      */
      private $trackAndTraceCarrier;
-
-    /**
-     * @var bool|null If TRUE, notify the customer of the status change. Also see 'note' and 'track_and_trace_code'.
-     */
-     private $mail;
 
     /**
      * @var string|null Delivery date in UTC
@@ -121,12 +116,17 @@ class Patch implements Model
      private $reservationnumber;
 
     /**
-     * @var \JacobDeKeizer\Ccv\Models\Entity\Personalinfo\Input|null Describes personal information of the customer
+     * @var \JacobDeKeizer\Ccv\Models\Internal\Entity\Personalinfo\Input Describes personal information of the customer
      */
      private $customer;
 
     /**
-     * @var \JacobDeKeizer\Ccv\Models\Entity\Orderrow\Input[]|null Describes a collection of order rows. Replaces all old order rows with the new orders rows. This will result in a recalculation of the order. If a product_id is provided the properties will be populated with product data. If you wish to overrule this data just add the property to the payload.
+     * @var int|null Unique id of the user (see GET /:version/users). The user's personal info will be added to this order. If additional information is provided in the property 'customer' that info will override this. If null, no user will be associated with this order.
+     */
+     private $userId;
+
+    /**
+     * @var \JacobDeKeizer\Ccv\Models\Internal\Entity\Orderrow\Input[] Describes a collection of order rows. If a product_id is provided the properties will be populated with product data. If you wish to overrule this data just add the property to the payload.
      */
      private $orderrows;
 
@@ -147,7 +147,15 @@ class Patch implements Model
     }
 
     /**
-     * @return bool|null If the order is marked as paid.
+     * @return bool|null If false, the order won't be marked as completed. The basket_href and checkout_href will be available for this resource. The consumer can be forwarded to these urls and complete the order. By default this value is TRUE.
+     */
+    public function getIsCompleted(): ?bool
+    {
+        return $this->isCompleted;
+    }
+
+    /**
+     * @return bool|null If the order is marked as paid
      */
     public function getPaid(): ?bool
     {
@@ -163,7 +171,7 @@ class Patch implements Model
     }
 
     /**
-     * @return float|null Total shipping costs. If this order already has a value for this field, it will be overwritten.
+     * @return float|null Total shipping costs
      */
     public function getTotalShipping(): ?float
     {
@@ -179,7 +187,7 @@ class Patch implements Model
     }
 
     /**
-     * @return float|null Extra costs added to the order, for instance as handling costs. This can be used as a discount less than zero. If this order already has a value for this property, it will be overwritten.
+     * @return float|null Extra costs added to the order, for instance as handling costs. This can be used as a discount less than zero.
      */
     public function getExtraCosts(): ?float
     {
@@ -187,7 +195,7 @@ class Patch implements Model
     }
 
     /**
-     * @return string|null Description of the extra costs. This is visible on the invoice and only if extra_costs is not zero. If extra_costs is supplied, it's advised to fill this property.
+     * @return string|null Description of the extra costs. This is visible on the invoice.
      */
     public function getExtraCostsDescription(): ?string
     {
@@ -203,7 +211,7 @@ class Patch implements Model
     }
 
     /**
-     * @return bool|null If order row prices contain taxes. Use this property to choose between an inc. VAT order and an ex. VAT order.
+     * @return bool|null If order row prices contain taxes. Use this field to choose between an inc. VAT order and an ex. VAT order.
      */
     public function getOrderRowTaxesIncluded(): ?bool
     {
@@ -211,7 +219,7 @@ class Patch implements Model
     }
 
     /**
-     * @return bool|null If shippingcosts are included in the total tax amount.
+     * @return bool|null If shippingcosts are included in the total tax amount
      */
     public function getShippingTaxesIncluded(): ?bool
     {
@@ -251,15 +259,7 @@ class Patch implements Model
     }
 
     /**
-     * @return string|null When the boolean 'mail' is set, the note will be appended to the email which is send to the customer.
-     */
-    public function getNote(): ?string
-    {
-        return $this->note;
-    }
-
-    /**
-     * @return string|null Track and Trace Code, supplied by shipping party. When the boolean 'mail' is set, the Track and Trace Code can be appended to the email which is send to the customer.
+     * @return string|null Track and Trace code, usually provided by the shipping service. If a note is emailed to the customer this value will be included.
      */
     public function getTrackAndTraceCode(): ?string
     {
@@ -272,14 +272,6 @@ class Patch implements Model
     public function getTrackAndTraceCarrier(): ?string
     {
         return $this->trackAndTraceCarrier;
-    }
-
-    /**
-     * @return bool|null If TRUE, notify the customer of the status change. Also see 'note' and 'track_and_trace_code'.
-     */
-    public function getMail(): ?bool
-    {
-        return $this->mail;
     }
 
     /**
@@ -315,17 +307,25 @@ class Patch implements Model
     }
 
     /**
-     * @return \JacobDeKeizer\Ccv\Models\Entity\Personalinfo\Input|null Describes personal information of the customer
+     * @return \JacobDeKeizer\Ccv\Models\Internal\Entity\Personalinfo\Input Describes personal information of the customer
      */
-    public function getCustomer(): ?\JacobDeKeizer\Ccv\Models\Entity\Personalinfo\Input
+    public function getCustomer(): \JacobDeKeizer\Ccv\Models\Internal\Entity\Personalinfo\Input
     {
         return $this->customer;
     }
 
     /**
-     * @return \JacobDeKeizer\Ccv\Models\Entity\Orderrow\Input[]|null Describes a collection of order rows. Replaces all old order rows with the new orders rows. This will result in a recalculation of the order. If a product_id is provided the properties will be populated with product data. If you wish to overrule this data just add the property to the payload.
+     * @return int|null Unique id of the user (see GET /:version/users). The user's personal info will be added to this order. If additional information is provided in the property 'customer' that info will override this. If null, no user will be associated with this order.
      */
-    public function getOrderrows(): ?array
+    public function getUserId(): ?int
+    {
+        return $this->userId;
+    }
+
+    /**
+     * @return \JacobDeKeizer\Ccv\Models\Internal\Entity\Orderrow\Input[] Describes a collection of order rows. If a product_id is provided the properties will be populated with product data. If you wish to overrule this data just add the property to the payload.
+     */
+    public function getOrderrows(): array
     {
         return $this->orderrows;
     }
@@ -342,7 +342,18 @@ class Patch implements Model
     }
 
     /**
-     * @param bool|null If the order is marked as paid.
+     * @param bool|null If false, the order won't be marked as completed. The basket_href and checkout_href will be available for this resource. The consumer can be forwarded to these urls and complete the order. By default this value is TRUE.
+     * @return self
+     */
+    public function setIsCompleted(?bool $isCompleted): self
+    {
+        $this->isCompleted = $isCompleted;
+        $this->propertyFilled('isCompleted');
+        return $this;
+    }
+
+    /**
+     * @param bool|null If the order is marked as paid
      * @return self
      */
     public function setPaid(?bool $paid): self
@@ -364,7 +375,7 @@ class Patch implements Model
     }
 
     /**
-     * @param float|null Total shipping costs. If this order already has a value for this field, it will be overwritten.
+     * @param float|null Total shipping costs
      * @return self
      */
     public function setTotalShipping(?float $totalShipping): self
@@ -386,7 +397,7 @@ class Patch implements Model
     }
 
     /**
-     * @param float|null Extra costs added to the order, for instance as handling costs. This can be used as a discount less than zero. If this order already has a value for this property, it will be overwritten.
+     * @param float|null Extra costs added to the order, for instance as handling costs. This can be used as a discount less than zero.
      * @return self
      */
     public function setExtraCosts(?float $extraCosts): self
@@ -397,7 +408,7 @@ class Patch implements Model
     }
 
     /**
-     * @param string|null Description of the extra costs. This is visible on the invoice and only if extra_costs is not zero. If extra_costs is supplied, it's advised to fill this property.
+     * @param string|null Description of the extra costs. This is visible on the invoice.
      * @return self
      */
     public function setExtraCostsDescription(?string $extraCostsDescription): self
@@ -419,7 +430,7 @@ class Patch implements Model
     }
 
     /**
-     * @param bool|null If order row prices contain taxes. Use this property to choose between an inc. VAT order and an ex. VAT order.
+     * @param bool|null If order row prices contain taxes. Use this field to choose between an inc. VAT order and an ex. VAT order.
      * @return self
      */
     public function setOrderRowTaxesIncluded(?bool $orderRowTaxesIncluded): self
@@ -430,7 +441,7 @@ class Patch implements Model
     }
 
     /**
-     * @param bool|null If shippingcosts are included in the total tax amount.
+     * @param bool|null If shippingcosts are included in the total tax amount
      * @return self
      */
     public function setShippingTaxesIncluded(?bool $shippingTaxesIncluded): self
@@ -485,18 +496,7 @@ class Patch implements Model
     }
 
     /**
-     * @param string|null When the boolean 'mail' is set, the note will be appended to the email which is send to the customer.
-     * @return self
-     */
-    public function setNote(?string $note): self
-    {
-        $this->note = $note;
-        $this->propertyFilled('note');
-        return $this;
-    }
-
-    /**
-     * @param string|null Track and Trace Code, supplied by shipping party. When the boolean 'mail' is set, the Track and Trace Code can be appended to the email which is send to the customer.
+     * @param string|null Track and Trace code, usually provided by the shipping service. If a note is emailed to the customer this value will be included.
      * @return self
      */
     public function setTrackAndTraceCode(?string $trackAndTraceCode): self
@@ -514,17 +514,6 @@ class Patch implements Model
     {
         $this->trackAndTraceCarrier = $trackAndTraceCarrier;
         $this->propertyFilled('trackAndTraceCarrier');
-        return $this;
-    }
-
-    /**
-     * @param bool|null If TRUE, notify the customer of the status change. Also see 'note' and 'track_and_trace_code'.
-     * @return self
-     */
-    public function setMail(?bool $mail): self
-    {
-        $this->mail = $mail;
-        $this->propertyFilled('mail');
         return $this;
     }
 
@@ -573,10 +562,10 @@ class Patch implements Model
     }
 
     /**
-     * @param \JacobDeKeizer\Ccv\Models\Entity\Personalinfo\Input|null Describes personal information of the customer
+     * @param \JacobDeKeizer\Ccv\Models\Internal\Entity\Personalinfo\Input Describes personal information of the customer
      * @return self
      */
-    public function setCustomer(?\JacobDeKeizer\Ccv\Models\Entity\Personalinfo\Input $customer): self
+    public function setCustomer(\JacobDeKeizer\Ccv\Models\Internal\Entity\Personalinfo\Input $customer): self
     {
         $this->customer = $customer;
         $this->propertyFilled('customer');
@@ -584,10 +573,21 @@ class Patch implements Model
     }
 
     /**
-     * @param \JacobDeKeizer\Ccv\Models\Entity\Orderrow\Input[]|null Describes a collection of order rows. Replaces all old order rows with the new orders rows. This will result in a recalculation of the order. If a product_id is provided the properties will be populated with product data. If you wish to overrule this data just add the property to the payload.
+     * @param int|null Unique id of the user (see GET /:version/users). The user's personal info will be added to this order. If additional information is provided in the property 'customer' that info will override this. If null, no user will be associated with this order.
      * @return self
      */
-    public function setOrderrows(?array $orderrows): self
+    public function setUserId(?int $userId): self
+    {
+        $this->userId = $userId;
+        $this->propertyFilled('userId');
+        return $this;
+    }
+
+    /**
+     * @param \JacobDeKeizer\Ccv\Models\Internal\Entity\Orderrow\Input[] Describes a collection of order rows. If a product_id is provided the properties will be populated with product data. If you wish to overrule this data just add the property to the payload.
+     * @return self
+     */
+    public function setOrderrows(array $orderrows): self
     {
         $this->orderrows = $orderrows;
         $this->propertyFilled('orderrows');
