+++ src/Models/Verto/Webshop/Resource/Ordernotes.php
@@ -1,17 +1,17 @@
 <?php
 
-namespace JacobDeKeizer\Ccv\Models\Webshop\Resource;
+namespace JacobDeKeizer\Ccv\Models\Verto\Webshop\Resource;
 
 use JacobDeKeizer\Ccv\Contracts\Model;
 use JacobDeKeizer\Ccv\Traits\FromArray;
 use JacobDeKeizer\Ccv\Traits\ToArray;
 
-class Orderaffiliatenetworks implements Model
+class Ordernotes implements Model
 {
     use FromArray, ToArray;
 
     /**
-     * @var string|null Link to affiliate of order
+     * @var string Link to notes added to this order
      */
      private $href;
 
@@ -24,18 +24,18 @@ class Orderaffiliatenetworks implements Model
     }
 
     /**
-     * @return string|null Link to affiliate of order
+     * @return string Link to notes added to this order
      */
-    public function getHref(): ?string
+    public function getHref(): string
     {
         return $this->href;
     }
 
     /**
-     * @param string|null Link to affiliate of order
+     * @param string Link to notes added to this order
      * @return self
      */
-    public function setHref(?string $href): self
+    public function setHref(string $href): self
     {
         $this->href = $href;
         $this->propertyFilled('href');
