+++ src/Models/Verto/Webshop/Resource/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace JacobDeKeizer\Ccv\Models\Webshop\Resource;
+namespace JacobDeKeizer\Ccv\Models\Verto\Webshop\Resource;
 
 use JacobDeKeizer\Ccv\Contracts\Model;
 use JacobDeKeizer\Ccv\Traits\FromArray;
